Extract fluid tower simulation module
